feat profil picture ranking & authAdminTeamGuard

+++ node-app/src/User/user.controller.ts
@@ -32,6 +32,7 @@ import { join } from 'path';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { Category } from 'src/Category/category.schema';
 import { Types } from 'mongoose';
+import { AdminTeamAuthGuard } from 'src/Auth/adminTeam';
 
 @Controller('users')
 export class UserController {
@@ -183,6 +184,19 @@ export class UserController {
     return { isAdminLoggedIn: true };
   }
 
+  @Get('check/adminTeam')
+  @UseGuards(AdminTeamAuthGuard, AuthenticatedGuard)
+  checkAdminTeamAuthentication(@Request() req) {
+    // if(req.user.isAdminTeam){
+    //   return { isAdminTeamLoggedIn: true };
+    // }
+    // else{
+    //   throw new HttpException('Unauthorized AdminTeam access', HttpStatus.FORBIDDEN);
+    // }
+    return { isAdminTeamLoggedIn: true };
+  }
+
+  @UseGuards(AdminAuthGuard)
   @Delete('delete/:userId')
   async deleteUser(
     @Param('userId') userId: string,
@@ -191,14 +205,11 @@ export class UserController {
       // Récupérer l'administrateur par défaut
       const defaultAdmin = await this.userService.findDefaultAdmin();
       const user = await this.userService.findById(userId);
-      // Vérifier si l'utilisateur à supprimer est l'administrateur par défaut
       if (user.email === defaultAdmin.email) {
         throw new Error(
           "Vous ne pouvez pas supprimer l'administrateur par défaut.",
         );
       }
-
-      // Supprimer l'utilisateur avec l'ID fourni
       await this.userService.deleteUser(userId);
       return { message: "L'utilisateur a été supprimé avec succès" };
     } catch (error) {
@@ -253,6 +264,60 @@ export class UserController {
   @Get('byId/:userId')
   async findById(@Param('userId') userId: string): Promise<User> {
     return this.userService.findById(userId);
+  }
+
+  @UseGuards(AuthenticatedGuard)
+  @Post('upload')
+  @UseInterceptors(
+    FileInterceptor('file', {
+      storage: diskStorage({
+        destination: './uploads', // Le répertoire où les fichiers seront stockés
+        filename: (req, file, cb) => {
+          const uniqueSuffix =
+            Date.now() + '-' + Math.round(Math.random() * 1e9);
+          // Utilisation de la fonction de rappel pour générer un nom de fichier unique
+          cb(null, file.originalname + '-' + uniqueSuffix);
+        },
+      }),
+    }),
+  )
+  async uploadProfilePicture(
+    @Request() req,
+    @Param('userId') userId: string,
+    @UploadedFile() file: Express.Multer.File,
+  ): Promise<User> {
+    console.log(file);
+    const user = req.user;
+    console.log(user);
+    return this.userService.updateProfilePicture(user.id, {
+      profilePicturePath: file.filename,
+    });
+    //return of({imagepath: file.filename});
+  }
+
+  @Get('profile-picture/:profilePicture')
+  FindProfilePicture(
+    @Param('profilePicture') profilePicture,
+    @Response() res,
+  ): Promise<User> {
+    return res.sendFile(join(process.cwd(), 'uploads/' + profilePicture));
+  }
+
+  @Put('update/:userId')
+  async updateUserProfile(
+    @Param('userId') userId: string,
+    @Body() updateUserDto: UpdateUserDto, // Créez un DTO approprié pour les données de mise à jour du profil
+  ): Promise<{ message: string }> {
+    try {
+      await this.userService.updateUserProfile(userId, updateUserDto); // Appelez votre service pour mettre à jour le profil de l'utilisateur
+      return { message: 'Profil utilisateur mis à jour avec succès' };
+    } catch (error) {
+      console.error(
+        "Erreur lors de la mise à jour du profil de l'utilisateur:",
+        error,
+      );
+      throw error;
+    }
   }
 
   @UseGuards(AuthenticatedGuard)
